requirements.txt
astroid==2.2.5
autopep8==1.4.4
backcall==0.2.0
certifi==2019.6.16
chardet==3.0.4
decorator==4.4.2
defusedxml==0.6.0
Django==2.2.13
django-address==0.2.5
django-address-app==0.1.7
django-allauth==0.39.1
django-appconf==1.0.4
django-cors-headers==3.0.2
django-credit-cards==0.4.1
django-imagekit==2.0.4
django-jet==1.0.8
django-phone-field==1.8.1
django-phonenumber-field==5.0.0
django-postgres-extensions==0.9.3
django-rest-auth==0.9.5
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
environs==9.3.1
gunicorn==19.9.0
idna==2.8
ipython==7.20.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.18.0
lazy-object-proxy==1.4.1
marshmallow==3.10.0
mccabe==0.6.1
mysqlclient==2.0.3
oauthlib==3.0.1
parso==0.8.1
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.2
prompt-toolkit==3.0.16
psycopg2==2.8.6
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pycodestyle==2.5.0
Pygments==2.8.0
PyJWT==1.7.1
pylint==2.3.1
python-dotenv==0.15.0
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
structlog==21.1.0
swapper==1.1.2.post1
traitlets==5.0.5
typed-ast==1.4.0
typing-extensions==3.7.4.3
urllib3==1.25.3
wcwidth==0.2.5
whitenoise==5.2.0
wrapt==1.11.2
